+++ fabric/src/main/java/club/iananderson/seasonhud/fabric/client/SeasonHudClientFabric.java
@@ -1,0 +1,18 @@
+package club.iananderson.seasonhud.fabric.client;
+
+import club.iananderson.seasonhud.Common;
+import club.iananderson.seasonhud.config.Config;
+import club.iananderson.seasonhud.fabric.event.ClientEvents;
+import fuzs.forgeconfigapiport.fabric.api.neoforge.v4.NeoForgeConfigRegistry;
+import net.fabricmc.api.ClientModInitializer;
+import net.neoforged.fml.config.ModConfig;
+
+public class SeasonHudClientFabric implements ClientModInitializer {
+
+  @Override
+  public void onInitializeClient() {
+    NeoForgeConfigRegistry.INSTANCE.register(Common.MOD_ID, ModConfig.Type.CLIENT, Config.GENERAL_SPEC,
+                                             "SeasonHUD-client.toml");
+    ClientEvents.register();
+  }
+}