+++ forge/src/main/java/club/iananderson/seasonhud/forge/client/overlays/SeasonHUDOverlay.java
@@ -1,0 +1,18 @@
+package club.iananderson.seasonhud.forge.client.overlays;
+
+import club.iananderson.seasonhud.client.overlays.SeasonHUDOverlayCommon;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.LayeredDraw;
+
+public class SeasonHUDOverlay implements LayeredDraw.Layer {
+  public static SeasonHUDOverlay HUD_INSTANCE;
+
+  public static void init() {
+    HUD_INSTANCE = new SeasonHUDOverlay();
+  }
+
+  @Override
+  public void render(GuiGraphics graphics, DeltaTracker deltaTracker) {
+    SeasonHUDOverlayCommon.render(graphics);
+  }
+}